+++ netket/__init__.py
@@ -61,4 +61,3 @@ from . import (
 # Main applications
 from .driver import VMC
 from .driver import SteadyState
-from .driver import QSR
