+++ netket/experimental/qsr/__init__.py
@@ -12,28 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-__all__ = [
-    "driver",
-    "dynamics",
-    "sampler",
-    "vqs",
-    "TDVP",
-    "hilbert",
-    "operator",
-    "logging",
-]
-
-from . import hilbert
-from . import operator
-from . import driver
-from . import dynamics
-from . import sampler
-from . import vqs
-from . import logging
-from . import qsr
-
-from .driver import TDVP
-from .qsr import QSR
+from .driver import QSR
 
 from netket.utils import _hide_submodules
