+++ netket/experimental/qsr/driver.py
@@ -13,455 +13,33 @@
 # limitations under the License.
 
 from typing import List, Optional, Tuple, Union
-from functools import partial
 
 import numpy as np
 import jax
 import jax.numpy as jnp
 
-from numba import njit
-
 from netket import jax as nkjax
 from netket.driver import AbstractVariationalDriver
 from netket.driver.vmc_common import info
-from netket.operator import AbstractOperator, LocalOperator
-from netket.hilbert import AbstractHilbert, Spin
+from netket.operator import AbstractOperator
 from netket.vqs import VariationalState
-from netket.vqs import ExactState
 from netket.optimizer import (
     identity_preconditioner,
     PreconditionerT,
 )
-from netket.utils import mpi
-from netket.utils.types import DType, Array
-from netket.utils.dispatch import dispatch
+from netket.utils.types import Array
 
 from netket.stats import statistics
 
+from .input_helpers import _convert_data, _compose_sampled_data
+from .logic_helpers import (
+    _grad_local_value_rotated,
+    _local_value_rotated_amplitude,
+    _compose_grads,
+    _grad_negative,
+)
+
 BaseType = Union[AbstractOperator, np.ndarray, str]
-
-
-def _build_rotation(
-    hi: Spin, basis: Union[List, str], dtype: Optional[DType] = np.complex64
-) -> AbstractOperator:
-    r"""
-    Construct basis rotation operators from a Pauli string of "X", "Y", "Z" and "I".
-
-    Args:
-        hi: The Hilbert space
-        basis: The Pauli string
-        dtype: The data type of the returned operator
-
-    Returns:
-        The rotation operator
-    """
-    localop = LocalOperator(hi, constant=1.0, dtype=dtype)
-    U_X = 1.0 / (np.sqrt(2)) * np.asarray([[1.0, 1.0], [1.0, -1.0]])
-    U_Y = 1.0 / (np.sqrt(2)) * np.asarray([[1.0, -1j], [1.0, 1j]])
-
-    assert len(basis) == hi.size
-
-    for (j, base) in enumerate(basis):
-        if base == "X":
-            localop *= LocalOperator(hi, U_X, [j])
-        elif base == "Y":
-            localop *= LocalOperator(hi, U_Y, [j])
-        elif base == "Z" or base == "I":
-            pass
-
-    return localop
-
-
-def _check_bases_type(Us: Union[List[BaseType], np.ndarray]) -> List[AbstractOperator]:
-    r"""
-    Check if the given bases are valid for the quantum state reconstruction driver.
-
-    Args:
-        Us (list or np.ndarray): A list of bases
-
-    Raises:
-        ValueError: When not given a list or np.ndarray
-        TypeError: If the type of the operators is not a child of AbstractOperator
-    """
-    if not (isinstance(Us, list) or isinstance(Us, np.ndarray)):
-        raise ValueError(
-            "The bases should be a list or np.ndarray(dtype=object)" " of the bases."
-        )
-
-    if isinstance(Us[0], AbstractOperator):
-        return Us
-
-    if isinstance(Us[0], str):
-        from netket.hilbert import Spin
-
-        hilbert = Spin(0.5, N=len(Us[0]))
-        N_samples = len(Us)
-
-        _cache = {}
-        _bases = np.empty(N_samples, dtype=object)
-
-        for (i, basis) in enumerate(Us):
-            if basis not in _cache:
-                U = _build_rotation(hilbert, basis)
-                _cache[basis] = U
-
-            _bases[i] = _cache[basis]
-        return _bases
-
-    raise TypeError("Unknown type of measurement basis.")
-
-
-def _convert_data(
-    sigma_s: np.ndarray,
-    Us: Union[List[BaseType], np.ndarray],
-    mixed_state_target: Optional[bool] = False,
-) -> Tuple[np.ndarray, np.ndarray, np.ndarray, int]:
-    r"""
-    Convert sampled states and rotation operators to a more direct computational format.
-    Specifically, for each sampled state sigma_s, find all the states sigma_p that have non-zero
-    matrix elements with sigma_s in the rotation operators Us. The corresponding
-    non-zero matrix elements (mels) and the indices of sigma_p and Us that divide
-    different sigma_s (secs) are also returned.
-
-    For pure states, we directly return sigma_p, mels and secs, so that one can use them
-    to compute <sigma_s|U|sigma_p> and sum over sigma_p. For mixed states, we return sigma_p x sigma_p,
-    mels and mels.conj() and the corresponding secs, so that one can use them to compute
-    <sigma_s|U|sigma_p><sigma_p'|U|sigma_s> and sum over both sigma_p and sigma_p'.
-
-    Args:
-        sigma_s (np.ndarray): The states
-        Us (np.ndarray or list): The list of rotations
-        mixed_state_target (bool): Whether to use mixed states or not
-
-    Returns:
-        sigma_p (np.ndarray): All the states that have non-zero matrix elements
-        with the input states sigma_s in the rotation operators Us.
-        mels (np.ndarray): The corresponding non-zero matrix elements.
-        secs (np.ndarray): Indices of sigma_p and Us that divide different sigma_s.
-        MAX_LEN (int): The maximum number of connected states.
-    """
-    Us = _check_bases_type(Us)
-
-    # Error message when user tries to convert less or more sigmas than Us
-    assert (
-        sigma_s.shape[0] == Us.shape[0]
-    ), "The number of samples should be equal to the number of rotations."
-
-    N = sigma_s.shape[-1]
-    sigma_s = sigma_s.reshape(-1, N)
-    Nb = sigma_s.shape[0]
-    N_target = N + N * mixed_state_target  # N or 2N if mixed state
-
-    # constant number of connected states per operator
-    Nc = Us[0].hilbert.local_size
-    sigma_p = np.zeros((0, N_target), dtype=sigma_s.dtype)
-    mels = np.zeros((0,), dtype=Us[0].dtype)
-    secs = np.zeros(Nb + 1, dtype=np.intp)
-    MAX_LEN = 0
-
-    last_i = 0
-    for (i, (sigma, U)) in enumerate(zip(sigma_s, Us)):
-        sigma_p_i, mels_i = U.get_conn(sigma)
-
-        if not mixed_state_target:
-            Nc = mels_i.size
-        else:
-            # size of the cartesian product sigma_p x sigma_p
-            Nc = mels_i.size**2
-
-        sigma_p = np.resize(sigma_p, (last_i + Nc, N_target))
-        mels = np.resize(mels, (last_i + Nc,))
-
-        if not mixed_state_target:
-            sigma_p[last_i:, :] = sigma_p_i
-            # <sigma_s|U|sigma_p>
-            mels[last_i:] = mels_i
-        else:
-            # indices of the cartesian product
-            x, y = np.meshgrid(np.arange(mels_i.size), np.arange(mels_i.size))
-            sigma_p[last_i:, :] = np.hstack(
-                [sigma_p_i[x.flatten()], sigma_p_i[y.flatten()]]
-            )
-            # <sigma_s|U|sigma_p><sigma_p'|U|sigma_s>
-            mels[last_i:] = np.prod(
-                np.stack(
-                    [mels_i[x.flatten()], np.conjugate(mels_i[y.flatten()])], axis=-1
-                ),
-                axis=-1,
-            )
-
-        secs[i] = last_i
-        last_i = last_i + Nc
-        MAX_LEN = max(Nc, MAX_LEN)
-
-    sigma_p = np.resize(sigma_p, (last_i + MAX_LEN, N_target))
-    mels = np.resize(mels, (last_i + MAX_LEN,))
-    sigma_p[last_i + Nc :, :] = 0.0
-    mels[last_i + Nc :] = 0.0
-    secs[-1] = last_i  # + MAX_LEN
-
-    return sigma_p, mels, secs, MAX_LEN
-
-
-@njit
-def _compose_sampled_data(
-    sigma_p: np.ndarray,
-    mels: np.ndarray,
-    secs: np.ndarray,
-    MAX_LEN: int,
-    sampled_indices: np.ndarray,
-    min_padding_factor: Optional[int] = 128,
-) -> Tuple[np.ndarray, np.ndarray, np.ndarray, int]:
-    r"""
-    Given the sampled indices, select the corresponding data from sigma_p, mels and secs.
-
-    Args:
-        sigma_p (np.ndarray): All the states that have non-zero matrix elements
-        with the input states sigma_s in the rotation operators Us.
-        mels (np.ndarray): The corresponding non-zero matrix elements.
-        secs (np.ndarray): Indices of sigma_p and Us that divide different sigma_s.
-        MAX_LEN (int): The maximum number of connected states.
-        sampled_indices (np.ndarray): The indices of the sampled states.
-        min_padding_factor (int): The minimum padding factor.
-
-    Returns:
-        The sampled sigma_p, mels, secs and MAX_LEN.
-    """
-    N_samples = sampled_indices.size
-    N = sigma_p.shape[-1]
-
-    _sigma_p = np.zeros((N_samples * MAX_LEN, N), dtype=sigma_p.dtype)
-    _mels = np.zeros((N_samples * MAX_LEN,), dtype=mels.dtype)
-    _secs = np.zeros((N_samples + 1,), dtype=secs.dtype)
-    _maxlen = 0
-
-    last_i = 0
-    for (n, i) in enumerate(sampled_indices):
-        start_i, end_i = secs[i], secs[i + 1]
-        len_i = end_i - start_i
-
-        # print(f"{n}, {i}, {last_i}, {start_i}, {end_i}")
-        _sigma_p[last_i : last_i + len_i, :] = sigma_p[start_i:end_i, :]
-        _mels[last_i : last_i + len_i] = mels[start_i:end_i]
-
-        last_i = last_i + len_i
-        _secs[n + 1] = last_i
-
-        _maxlen = max(_maxlen, len_i)
-
-    padding_factor = max(
-        MAX_LEN, min_padding_factor
-    )  # minimum padding at 64 to avoid excessive recompilation
-    padded_size = padding_factor * int(np.ceil(last_i / padding_factor))
-
-    _sigma_p = _sigma_p[:padded_size, :]
-    _mels = _mels[:padded_size]
-
-    return _sigma_p, _mels, _secs, _maxlen
-
-
-@partial(jax.jit, static_argnums=(0))
-def _avg_O(afun, pars, model_state, sigma):
-    r"""
-    Compute the gradient of the normalization log Z by sample mean.
-    For example, for an unnormalized probability distribution p_i, we have
-    :math:`\nabla log Z = \frac{1}{Z} \sum_i \nabla p_i = \sum_i \frac{1}{Z} p_i \nabla \log p_i \approx \frac{1}{N} \sum_i \nabla \log p_i`
-
-    Args:
-        afun (function): The unnormalized log probability function.
-        pars (PyTree): The parameters of the model.
-        model_state (PyTree): The model state.
-        sigma (np.ndarray): The sampled states.
-
-    Returns:
-        The gradient of the normalization log Z.
-    """
-    sigma = sigma.reshape((-1, sigma.shape[-1]))
-    _, vjp = nkjax.vjp(lambda W: afun({"params": W, **model_state}, sigma), pars)
-    (O_avg,) = vjp(jnp.ones(sigma.shape[0]) / sigma.shape[0])
-    return jax.tree_map(lambda x: mpi.mpi_mean_jax(x)[0], O_avg)
-
-
-@dispatch
-def _grad_negative(state_diag):
-    r"""
-    Compute the gradient of the normalization log Z by sample mean.
-    For example, for an unnormalized probability distribution p_i, we have
-    :math:`\nabla log Z = \frac{1}{Z} \sum_i \nabla p_i = \sum_i \frac{1}{Z} p_i \nabla \log p_i \approx \frac{1}{N} \sum_i \nabla \log p_i`
-
-    Args:
-        state_diag (VariationalState): The variational state.
-
-    Returns:
-        The gradient of the normalization log Z.
-    """
-    return _avg_O(
-        state_diag._apply_fun,
-        state_diag.parameters,
-        state_diag.model_state,
-        state_diag.samples,
-    )
-
-
-@partial(jax.jit, static_argnums=(0, 1))
-def _avg_O_exact(hilbert: AbstractHilbert, afun, pars, model_state):
-    r"""
-    Same as _avg_O, but for ExactState.
-    """
-    sigma = hilbert.all_states()
-    sigma = sigma.reshape((-1, sigma.shape[-1]))
-    _, vjp = nkjax.vjp(lambda W: afun({"params": W, **model_state}, sigma), pars)
-    psi_2 = jnp.abs(jnp.exp(afun({"params": pars, **model_state}, sigma))) ** 2
-    psi_2 /= jnp.sum(psi_2)
-    (O_avg,) = vjp(psi_2)
-    return jax.tree_map(lambda x: mpi.mpi_mean_jax(x)[0], O_avg)
-
-
-@dispatch
-def _grad_negative(state_diag: ExactState):  # noqa: F811
-    r"""
-    Same as _grad_negative, but for ExactState.
-    """
-    return _avg_O_exact(
-        state_diag._hilbert,
-        state_diag._apply_fun,
-        state_diag.parameters,
-        state_diag.model_state,
-    )
-
-
-####
-
-
-def _sum_sections(arr: Array, secs: Array) -> Array:
-    """
-    Sum the elements of arr in sections defined by secs.
-    Equivalent to
-    for i in range(secs.size-1):
-        out[i] = jnp.sum(arr[secs[i]:secs[i+1]])
-
-    secs must be 1 longer than the number of sections desired.
-    MAX_SIZE is the max length of a section
-
-    Args:
-        arr (np.ndarray): The array to be summed.
-        secs (np.ndarray): The indices that define the sections.
-
-    Returns:
-        The sum of the array in sections.
-    """
-    offsets = secs[:-1]
-    sizes = secs[1:] - offsets
-
-    N_rows = secs.size - 1
-    # Construct the indices necessary to perform the segment_sum
-    indices = jnp.repeat(jnp.arange(N_rows), sizes, total_repeat_length=arr.size)
-
-    return jax.ops.segment_sum(
-        arr, indices, num_segments=N_rows, indices_are_sorted=True
-    )
-
-
-@partial(jax.jit, static_argnums=(0))
-def _local_value_rotated_kernel(log_psi, pars, sigma_p, mel, secs):
-    r"""
-    Compute the log probability amplitude \log <sigma_p|U|psi> of obtaining an outcome state sigma_p in the rotated basis.
-    For mixed states, it's \log <sigma_p|U \rho U^\dagger|sigma_p>.
-
-    Args:
-        log_psi (function): The log wavefunction or density matrix.
-        pars (PyTree): The parameters of the model.
-        sigma_p (np.ndarray): The sampled states.
-        mel (np.ndarray): The matrix elements of the rotations.
-        secs (np.ndarray): The indices that define the connected sections of different basis.
-
-    Returns:
-        The probability amplitude of obtaining an outcome state sigma_p in the rotated basis.
-    """
-    log_psi_sigma_p = log_psi(pars, sigma_p)
-    U_sigma_sigma_p_psi_sigma_p = mel * jnp.exp(log_psi_sigma_p)
-
-    return jnp.log(_sum_sections(U_sigma_sigma_p_psi_sigma_p, secs))
-
-
-@partial(jax.jit, static_argnums=(0))
-def _grad_local_value_rotated(log_psi, pars, model_state, sigma_p, mel, secs):
-    r"""
-    Compute the gradient of the log probability amplitude \log <sigma_p|U|psi> of obtaining an outcome state sigma_p in the rotated basis.
-    For mixed states, it's the gradient of \log <sigma_p|U \rho U^\dagger|sigma_p>.
-
-    Args:
-        log_psi (function): The log wavefunction or density matrix.
-        pars (PyTree): The parameters of the model.
-        model_state (PyTree): The model state.
-        sigma_p (np.ndarray): The sampled states.
-        mel (np.ndarray): The matrix elements of the rotations.
-        secs (np.ndarray): The indices that define the connected sections of different basis.
-
-    Returns:
-        The gradient of the probability amplitude of obtaining an outcome state sigma_p in the rotated basis.
-    """
-    log_val_rotated, vjp = nkjax.vjp(
-        lambda W: _local_value_rotated_kernel(
-            log_psi, {"params": W, **model_state}, sigma_p, mel, secs
-        ),
-        pars,
-    )
-    log_val_rotated, _ = mpi.mpi_mean_jax(log_val_rotated)
-
-    (O_avg,) = vjp(jnp.ones_like(log_val_rotated) / log_val_rotated.size)
-
-    O_avg = jax.tree_map(lambda x: mpi.mpi_mean_jax(x)[0], O_avg)
-
-    return log_val_rotated, O_avg
-
-
-@jax.jit
-def _compose_grads(grad_neg, grad_pos):
-    r"""
-    Compose the negative gradient (the normalization log Z) and the positive gradient (log probability amplitude).
-    For complex parameters, the complex conjugate is taken.
-
-    Args:
-        grad_neg (PyTree): The negative gradient.
-        grad_pos (PyTree): The positive gradient.
-
-    Returns:
-        The composed gradient.
-    """
-    return jax.tree_util.tree_map(
-        lambda x, y: jnp.conj(x - y),
-        grad_neg,
-        grad_pos,
-    )
-
-
-# for nll
-@partial(jax.jit, static_argnums=(0,))
-def _local_value_rotated_amplitude(log_psi, pars, sigma_p, mel, secs):
-    r"""
-    Only for monitoring negative log likelihood.
-
-    Compute the log probability amplitude squared \log |<sigma_p|U|psi>|^2 of obtaining an outcome state sigma_p in the rotated basis.
-    For mixed states, it's <sigma_p|U \rho U^\dagger|sigma_p>.
-
-    Args:
-        log_psi (function): The log wavefunction or density matrix.
-        pars (PyTree): The parameters of the model.
-        sigma_p (np.ndarray): The sampled states.
-        mel (np.ndarray): The matrix elements of the rotations.
-        secs (np.ndarray): The indices that define the connected sections of different basis.
-
-    Returns:
-        The probability amplitude of obtaining an outcome state sigma_p in the rotated basis.
-    """
-    log_psi_sigma_p = log_psi(pars, sigma_p)
-    U_sigma_sigma_p_psi_sigma_p = mel * jnp.exp(log_psi_sigma_p)
-
-    return jnp.log(jnp.abs(_sum_sections(U_sigma_sigma_p_psi_sigma_p, secs)) ** 2)
-
-
-####
 
 
 class QSR(AbstractVariationalDriver):
